backend/requirements.txt
fastapi
uvicorn
motor
pydantic>=2.0
pydantic-settings
python-jose
passlib==1.7.4
bcrypt==4.0.1
python-dotenv
